feat: force layer values to be collections
Callable layer values are no longer interpreted as functions capable of
producing values.  Instead, they are treated as values that should be
loaded just-in-time.  This is consistent with the behavior of layer
locations, and makes it easier to make layer values that depend on the
attributes of the object in question (which typically won't exist when
layers are being loaded).

Fixes #14