fix: use 'is' when comparing values to *ignore*.
Previously I used '==', but that can raise depending on how `__eq__()`
is implemented.  For example:

  class MyCustomClass:

      def __init__(self, x, y):
          self.x = x
          self.y = y

      def __eq__(self, other):
        return self.x == other.x and self.y == other.y

  obj = MyCustomClass(1, 2):
  obj == None

One could argue that this is a poor implementation of `__eq__()`, but I
think it makes sense to implement this feature with an operator that
can't raise.  In addition, I can't really think of a scenario where the
ignore value would be something other than None or an `object()`
sentinel, and in both cases 'is' is the most correct comparison to make.