feat: rename 'attr' back to 'param'
After thinking about it more, I decided that "param" is a more specific
and more corect term for these descriptors.  In computer science, a
parameter is a name/placeholder for a value that will be provided by the
user.  That's exactly what these descriptors are, and in this case the
values are filled in from various sources.  Of course these descriptors
also behave as attributes, but that's a more general termm.

Additionally, the term "param" is more unique than "attr", and so is
easier to refer to in documentation and such.